Add support for skipping CMake invocation in nix-shell

+++ chat/default.nix
@@ -56,33 +56,35 @@ in stdenv.mkDerivation rec {
       cd ${toString src}
       [[ -d "$(pwd)/build" ]] && rm -rf build
       source ${../helpers}
-      OLD_CMAKE_FLAGS="$cmakeFlags"
-      cmakeFlags="-DCMAKE_C_FLAGS=-v -DCMAKE_CXX_FLAGS=-v $cmakeFlags" \
-          cmakeConfigurePhase
-      BUILD_LOG="$(mktemp -p /tmp build-log.XXXXXX)"
-      ninja &> "$BUILD_LOG"
-      echo ""
-      echo ""
-      echo ""
-      note "FLYCHECK-CLANG-INCLUDE-PATH:"
-      echo ""
-      INCLUDE_PATH="$(generate-emacs-include-path "$BUILD_LOG")"
-      echo "$INCLUDE_PATH"
-      echo ""
-      echo ""
-      echo ""
-      read -p "Copy generated include path into clipboard? [y/N] "
-      if text-represents-yes "$REPLY"; then
-          echo "$INCLUDE_PATH" | clipboard
+      if [[ -z "$CHAT_NIX_SKIP_CMAKE" ]]; then
+          OLD_CMAKE_FLAGS="$cmakeFlags"
+          cmakeFlags="-DCMAKE_C_FLAGS=-v -DCMAKE_CXX_FLAGS=-v $cmakeFlags" \
+              cmakeConfigurePhase
+          BUILD_LOG="$(mktemp -p /tmp build-log.XXXXXX)"
+          ninja &> "$BUILD_LOG"
+          echo ""
+          echo ""
+          echo ""
+          note "FLYCHECK-CLANG-INCLUDE-PATH:"
+          echo ""
+          INCLUDE_PATH="$(generate-emacs-include-path "$BUILD_LOG")"
+          echo "$INCLUDE_PATH"
+          echo ""
+          echo ""
+          echo ""
+          read -p "Copy generated include path into clipboard? [y/N] "
+          if text-represents-yes "$REPLY"; then
+              echo "$INCLUDE_PATH" | clipboard
+          fi
+          echo ""
+          echo ""
+          echo ""
+          rm "$BUILD_LOG"
+          cmakeFlags="$OLD_CMAKE_FLAGS"
+          unset OLD_CMAKE_FLAGS BUILD_LOG
+          refresh-plugin-build "${toString src}"
       fi
-      echo ""
-      echo ""
-      echo ""
-      rm "$BUILD_LOG"
-      cmakeFlags="$OLD_CMAKE_FLAGS"
-      unset OLD_CMAKE_FLAGS BUILD_LOG
-      refresh-plugin-build "${toString src}"
-      echo for qtcreator: ${allHeaders}
+      echo "for qtcreator: ${allHeaders}"
       note "Entering nix-shell"
   '';
 }
